Moving out code to handle single item requests in a new file.
